v1.14.3
Fix issue where addons from TukUI would not install correctly.
Add confirmation box to addon uninstall.
Newly installed addons should appear in the correct spot in the list, not the bottom.
Uninstalling an addon sould be reflected on the "Get Addons" page.
Refreshing the "Get Addons" page should show the correct addons now.
Addon lists should scroll a little more smoothly.

+++ WowUp.WPF/ViewModels/AddonListItemViewModel.cs
@@ -294,11 +294,11 @@ namespace WowUp.WPF.ViewModels
 
             IsAutoUpdated = _addon.AutoUpdateEnabled;
 
-            if(ChannelType == AddonChannelType.Beta)
+            if (ChannelType == AddonChannelType.Beta)
             {
                 ChannelNameBrush = (SolidColorBrush)new BrushConverter().ConvertFrom("#0070DD");
             }
-            else if(ChannelType == AddonChannelType.Alpha)
+            else if (ChannelType == AddonChannelType.Alpha)
             {
                 ChannelNameBrush = (SolidColorBrush)new BrushConverter().ConvertFrom("#A335EE");
             }
@@ -357,6 +357,16 @@ namespace WowUp.WPF.ViewModels
 
         private async void OnUninstall()
         {
+            var messageBoxResult = MessageBox.Show(
+                $"Are you sure you want to remove {Addon.Name}? This will remove all related folders from your World of Warcraft folder.",
+                "Uninstall Addon?",
+                MessageBoxButton.YesNo);
+
+            if (messageBoxResult != MessageBoxResult.Yes)
+            {
+                return;
+            }
+
             try
             {
                 await _addonService.UninstallAddon(Addon);
